requirements.txt
streamlit
google-genai
requests
pdfplumber
python-docx
